+++ source/ada/lsp-ada_completions-keywords.ads
@@ -1,7 +1,7 @@
 ------------------------------------------------------------------------------
 --                         Language Server Protocol                         --
 --                                                                          --
---                     Copyright (C) 2018-2021, AdaCore                     --
+--                     Copyright (C) 2018-2026, AdaCore                     --
 --                                                                          --
 -- This is free software;  you can redistribute it  and/or modify it  under --
 -- terms of the  GNU General Public License as published  by the Free Soft- --
@@ -16,6 +16,8 @@
 ------------------------------------------------------------------------------
 --  A completion provider for keywords
 
+with LSP.Ada_Completions.Filters;
+
 package LSP.Ada_Completions.Keywords is
 
    type Keyword_Completion_Provider is new Completion_Provider with
@@ -27,8 +29,7 @@ package LSP.Ada_Completions.Keywords is
       Token  : Libadalang.Common.Token_Reference;
       Node   : Libadalang.Analysis.Ada_Node;
       Filter : in out LSP.Ada_Completions.Filters.Filter;
-      Names  : in out Ada_Completions.Completion_Maps.Map;
-      Result : in out LSP.Structures.CompletionList);
+      Result : out Ada_Completions.Completion_Result);
    --  Get completion for keywords, filtering them with the prefix.
 
 end LSP.Ada_Completions.Keywords;